requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.0
pystray>=0.19.4
Pillow>=10.0.0
pywin32>=306
deepfilternet>=0.5.6      # optional: needs Rust + MSVC to build
pyrnnoise>=0.4.0          # AI noise suppression, no compilation needed
